requirements.txt
Django>=2.1,<2.2
Pillow>=5.2,<5.3
requests>=2.19.0,<2.20.0
uWSGI==2.0.17.1
psycopg2-binary==2.7.5
django-debug-toolbar==1.9.1
django-admin-sortable2==0.6.21
django-allauth==0.38.0
django-reversion==3.0.0
gunicorn>=19.9,<19.10
ipdb==0.11

djangorestframework==3.8.2
django-filter==2.0.0
flex==6.13.2
drf-yasg==1.10.0
# This has a fix for migrations error
git+git://github.com/aschn/drf-tracking.git

django-enumfields==0.10.1
celery[redis]==4.2.1
django-redis==4.9.0
# Workaround for https://github.com/niwinz/django-redis/issues/342
redis==2.10.6

metadata_parser==0.9.21
beautifulsoup4==4.6.3
lxml==4.2.5

# This unreleased version has task name and other parameters
git+git://github.com/celery/django-celery-results.git
django-ipware==2.1.0
raven==6.9.0
# Version in pip doesn't have EXIF bugfix
git+git://github.com/un1t/django-resized.git
django-crispy-forms==1.7.2
django-recaptcha3==0.0.1
python-dateutil==2.7.3

sendgrid==5.6.0
django-sendgrid-v5==0.7.0